Modify how strings are packed in the RPCs, Maximum string size     increased from 64KB (16-bit size field) to 4GB (32-bit size field).

+++ src/common/io_hdr.c
@@ -179,7 +179,7 @@ io_init_msg_packed_size(void)
 
 	len = sizeof(uint16_t)        /* version */
 		+ sizeof(uint32_t)    /* nodeid */
-		+ (SLURM_IO_KEY_SIZE + sizeof(uint16_t))  /* signature */
+		+ (SLURM_IO_KEY_SIZE + sizeof(uint32_t))  /* signature */
 		+ sizeof(uint32_t)    /* stdout_objs */
 		+ sizeof(uint32_t);   /* stderr_objs */
 	return len;
@@ -193,15 +193,14 @@ io_init_msg_pack(struct slurm_io_init_msg *hdr, Buf buffer)
 	pack32(hdr->stdout_objs, buffer);
 	pack32(hdr->stderr_objs, buffer);
 	packmem((char *) hdr->cred_signature,
-		(uint16_t) SLURM_IO_KEY_SIZE, buffer);
+		(uint32_t) SLURM_IO_KEY_SIZE, buffer);
 }
 
 
 static int
 io_init_msg_unpack(struct slurm_io_init_msg *hdr, Buf buffer)
 {
-	uint16_t val;
-
+	uint32_t val;
 	safe_unpack16(&hdr->version, buffer);
 	safe_unpack32(&hdr->nodeid, buffer);
 	safe_unpack32(&hdr->stdout_objs, buffer);
